+++ packages/general/src/crypto/Crypto.ts
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { Diagnostic } from "#log/Diagnostic.js";
+import { Logger } from "#log/Logger.js";
 import { Boot } from "#util/Boot.js";
 import { MaybePromise } from "#util/Promises.js";
 import * as mod from "@noble/curves/abstract/modular";
@@ -29,6 +31,8 @@ export const CRYPTO_AUTH_TAG_LENGTH = 16;
 export const CRYPTO_SYMMETRIC_KEY_LENGTH = 16;
 export type CryptoDsaEncoding = "ieee-p1363" | "der";
 
+const logger = Logger.get("Crypto");
+
 /**
  * These are the cryptographic primitives required to implement the Matter protocol.
  *
@@ -41,6 +45,11 @@ export type CryptoDsaEncoding = "ieee-p1363" | "der";
  * WARNING: The standard implementation is unaudited.  See relevant warnings in StandardCrypto.ts.
  */
 export interface Crypto {
+    /**
+     * The name used in log messages.
+     */
+    implementationName: string;
+
     /**
      * Encrypt using AES-CCM with constants limited to those required by Matter.
      */
@@ -111,40 +120,80 @@ export interface Crypto {
     generateDhSecret(key: PrivateKey, peerKey: PublicKey): MaybePromise<Uint8Array>;
 }
 
+let logImplementationName = true;
+let defaultInstance: undefined | Crypto;
+let defaultProvider: undefined | (() => Crypto);
+
 /**
  * Crypto support functions.
  */
 export const Crypto = {
-    get(): Crypto {
-        throw new NoProviderError("No provider configured");
+    /**
+     * The default crypto implementation.
+     */
+    get default() {
+        if (defaultInstance) {
+            return defaultInstance;
+        }
+
+        if (defaultProvider === undefined) {
+            throw new NoProviderError("There is no cryptography implementation installed");
+        }
+
+        defaultInstance = defaultProvider();
+
+        if (logImplementationName) {
+            logger.debug("Using", Diagnostic.strong(defaultInstance.implementationName), "cryptography implementation");
+        }
+
+        return defaultInstance;
+    },
+
+    get provider(): undefined | (() => Crypto) {
+        return defaultProvider;
+    },
+
+    /**
+     * Set the default crypto provider.
+     */
+    set provider(provider: () => Crypto) {
+        if (defaultProvider === provider) {
+            return;
+        }
+        defaultProvider = undefined;
+        defaultProvider = provider;
+    },
+
+    get implementationName() {
+        return Crypto.default.implementationName;
     },
 
     encrypt(key: Uint8Array, data: Uint8Array, nonce: Uint8Array, aad?: Uint8Array) {
-        return Crypto.get().encrypt(key, data, nonce, aad);
+        return Crypto.default.encrypt(key, data, nonce, aad);
     },
 
     decrypt(key: Uint8Array, data: Uint8Array, nonce: Uint8Array, aad?: Uint8Array) {
-        return Crypto.get().decrypt(key, data, nonce, aad);
+        return Crypto.default.decrypt(key, data, nonce, aad);
     },
 
     getRandomData(length: number) {
-        return Crypto.get().getRandomData(length);
+        return Crypto.default.getRandomData(length);
     },
 
     getRandom() {
-        return Crypto.get().getRandomData(CRYPTO_RANDOM_LENGTH);
+        return Crypto.default.getRandomData(CRYPTO_RANDOM_LENGTH);
     },
 
     getRandomUInt16() {
-        return new DataReader(Crypto.get().getRandomData(2), Endian.Little).readUInt16();
+        return new DataReader(Crypto.default.getRandomData(2), Endian.Little).readUInt16();
     },
 
     getRandomUInt32() {
-        return new DataReader(Crypto.get().getRandomData(4), Endian.Little).readUInt32();
+        return new DataReader(Crypto.default.getRandomData(4), Endian.Little).readUInt32();
     },
 
     getRandomBigUInt64() {
-        return new DataReader(Crypto.get().getRandomData(8), Endian.Little).readUInt64();
+        return new DataReader(Crypto.default.getRandomData(8), Endian.Little).readUInt64();
     },
 
     getRandomBigInt(size: number, maxValue?: bigint) {
@@ -159,47 +208,51 @@ export const Crypto = {
     },
 
     computeSha256(data: Uint8Array | Uint8Array[]) {
-        return Crypto.get().computeSha256(data);
+        return Crypto.default.computeSha256(data);
     },
 
     createPbkdf2Key(secret: Uint8Array, salt: Uint8Array, iteration: number, keyLength: number) {
-        return Crypto.get().createPbkdf2Key(secret, salt, iteration, keyLength);
+        return Crypto.default.createPbkdf2Key(secret, salt, iteration, keyLength);
     },
 
     createHkdfKey(secret: Uint8Array, salt: Uint8Array, info: Uint8Array, length?: number) {
-        return Crypto.get().createHkdfKey(secret, salt, info, length);
+        return Crypto.default.createHkdfKey(secret, salt, info, length);
     },
 
     signHmac(key: Uint8Array, data: Uint8Array) {
-        return Crypto.get().signHmac(key, data);
+        return Crypto.default.signHmac(key, data);
     },
 
     signEcdsa(privateKey: JsonWebKey, data: Uint8Array | Uint8Array[], dsaEncoding?: CryptoDsaEncoding) {
-        return Crypto.get().signEcdsa(privateKey, data, dsaEncoding);
+        return Crypto.default.signEcdsa(privateKey, data, dsaEncoding);
     },
 
     verifyEcdsa(publicKey: JsonWebKey, data: Uint8Array, signature: Uint8Array, dsaEncoding?: CryptoDsaEncoding) {
-        return Crypto.get().verifyEcdsa(publicKey, data, signature, dsaEncoding);
+        return Crypto.default.verifyEcdsa(publicKey, data, signature, dsaEncoding);
     },
 
     createKeyPair() {
-        return Crypto.get().createKeyPair();
+        return Crypto.default.createKeyPair();
     },
 
     generateDhSecret(key: PrivateKey, peerKey: PublicKey) {
-        return Crypto.get().generateDhSecret(key, peerKey);
+        return Crypto.default.generateDhSecret(key, peerKey);
     },
 };
 
 Crypto satisfies Crypto;
 
 Boot.init(() => {
-    Crypto.get = () => {
-        throw new NoProviderError("No provider configured");
-    };
+    logImplementationName = true;
+    defaultInstance = undefined;
+    defaultProvider = undefined;
 
-    // Hook for testing frameworks
+    // Testing framework configuration
     if (typeof MatterHooks !== "undefined") {
+        // Crypto access occurs before log messages are intercepted so do not log implementation in test environment
+        logImplementationName = true;
+
+        // Configure mocking
         MatterHooks.cryptoSetup?.(Crypto);
     }
 });